+++ merged-packages/bitcoin-wallet-snap/src/handlers/KeyringHandler.ts
@@ -13,6 +13,7 @@ import {
   ListAccountsRequestStruct,
   ListAccountTransactionsRequestStruct,
   MetaMaskOptionsStruct,
+  ResolveAccountAddressRequestStruct,
   SetSelectedAccountsRequestStruct,
   SubmitRequestRequestStruct,
 } from '@metamask/keyring-api';
@@ -29,8 +30,9 @@ import type {
   MetaMaskOptions,
   DiscoveredAccount,
   KeyringRequest,
+  ResolvedAccountAddress,
 } from '@metamask/keyring-api';
-import type { Json, JsonRpcRequest } from '@metamask/snaps-sdk';
+import type { CaipChainId, Json, JsonRpcRequest } from '@metamask/snaps-sdk';
 import {
   assert,
   boolean,
@@ -54,6 +56,7 @@ import {
   caipToAddressType,
   scopeToNetwork,
   networkToScope,
+  NetworkStruct,
 } from './caip';
 import { CronMethod } from './CronHandler';
 import type { KeyringRequestHandler } from './KeyringRequestHandler';
@@ -62,7 +65,7 @@ import {
   mapToKeyringAccount,
   mapToTransaction,
 } from './mappings';
-import { validateSelectedAccounts } from './validation';
+import { BtcWalletRequestStruct, validateSelectedAccounts } from './validation';
 import type { AccountUseCases } from '../use-cases/AccountUseCases';
 import { runSnapActionSafely } from '../utils/snapHelpers';
 
@@ -159,6 +162,13 @@ export class KeyringHandler implements Keyring {
         assert(request, SetSelectedAccountsRequestStruct);
         await this.setSelectedAccounts(request.params.accounts);
         return null;
+      }
+      case `${KeyringRpcMethod.ResolveAccountAddress}`: {
+        assert(request, ResolveAccountAddressRequestStruct);
+        return this.resolveAccountAddress(
+          request.params.scope,
+          request.params.request,
+        );
       }
 
       default: {
@@ -384,6 +394,57 @@ export class KeyringHandler implements Keyring {
       method: CronMethod.SyncSelectedAccounts,
       params: { accountIds: accounts },
     });
+  }
+
+  /**
+   * Resolves the address of an account from a signing request.
+   *
+   * This is required by the routing system of MetaMask to dispatch
+   * incoming non-EVM dapp signing requests.
+   *
+   * @param scope - Request's scope (CAIP-2).
+   * @param request - Signing request object.
+   * @returns A Promise that resolves to the account address that must
+   * be used to process this signing request, or null if none candidates
+   * could be found.
+   */
+  async resolveAccountAddress(
+    scope: CaipChainId,
+    request: JsonRpcRequest,
+  ): Promise<ResolvedAccountAddress | null> {
+    try {
+      assert(scope, NetworkStruct);
+      const { method, params } = request;
+
+      const requestWithoutCommonHeader = { method, params };
+      assert(requestWithoutCommonHeader, BtcWalletRequestStruct);
+
+      const allAccounts = await this.listAccounts();
+
+      const accountsWithThisScope = allAccounts.filter((account) =>
+        account.scopes.includes(scope),
+      );
+
+      if (accountsWithThisScope.length === 0) {
+        throw new Error('No accounts with this scope');
+      }
+
+      const { address: addressToValidate } =
+        requestWithoutCommonHeader.params.account;
+
+      const foundAccount = accountsWithThisScope.find(
+        (account) => account.address === addressToValidate,
+      );
+
+      if (!foundAccount) {
+        throw new Error('Account not found');
+      }
+
+      return { address: `${scope}:${addressToValidate}` };
+    } catch (error: unknown) {
+      this.#logger.error({ error }, 'Error resolving account address');
+      return null;
+    }
   }
 
   #extractAddressType(path: string): AddressType {